[PIB-51] Split IoT bounded context in own folder
