+++ backend/PIB.Domain/IoT/Actuators/GrowthLight/GrowthLightConfig.cs
@@ -1,4 +1,4 @@
-namespace Domain.Actuators.Light;
+namespace Domain.IoT.Actuators.GrowthLight;
 
 public class GrowthLightConfig
 {
