+++ backend/PIB.Domain/IoT/Sensors/ISensor.cs
@@ -1,4 +1,4 @@
-namespace Domain.Sensors;
+namespace Domain.IoT.Sensors;
 
 public interface ISensor
 {
