+++ backend/PIB.Domain/IoT/Sensors/SoilMoisture/SoilMoistureData.cs
@@ -1,4 +1,4 @@
-namespace Domain.Sensors.SoilMoisture;
+namespace Domain.IoT.Sensors.SoilMoisture;
 
 public class SoilMoistureData: ISensorData
 {
